feat: CoreDataStack 생성
