+++ Diary/Resource/Model/Diary+CoreDataClass.swift
@@ -10,6 +10,4 @@ import Foundation
 import CoreData
 
 @objc(Diary)
-public class Diary: NSManagedObject {
-
-}
+public class Diary: NSManagedObject { }
