+++ Diary/Resource/Model/Diary+CoreDataProperties.swift
@@ -18,5 +18,4 @@ extension Diary {
     @NSManaged public var title: String?
     @NSManaged public var body: String?
     @NSManaged public var createdAt: Date?
-
 }
